Dog Breed project: Steps 6 and 7 finished
Added the functionality to work with the Xception model.
Added the function to determine if the image is a human, a dog or anything else
Added 12 images to test the algorithm.
Added the code in the last cell that run the function to test the predictions for each image automatically